requirements.txt
flask
Pillow
google-generativeai
